# backend/swagger/openApi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/register": {
      "post": {
        "summary": "Register a new user",
        "operationId": "registerUser",
        "tags": ["Users"],
        "requestBody": {
          "description": "User registration data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tokens"
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "User login",
        "operationId": "loginUser",
        "tags": ["Users"],
        "requestBody": {
          "description": "User login data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tokens"
                }
              }
            }
          }
        }
      }
    },
    "/api/refresh-token": {
      "post": {
        "summary": "Refresh Access Token",
        "description": "Refresh the access token using a valid refresh token.",
        "operationId": "refreshToken",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": ["refreshToken"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful token refresh",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "The newly generated access token."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todo/list": {
      "get": {
        "summary": "Get the list of todos",
        "operationId": "getTodoList",
        "tags": ["Todos"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todo/{id}": {
      "get": {
        "summary": "Get a specific todo by ID",
        "operationId": "getTodoById",
        "tags": ["Todos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a todo by ID",
        "operationId": "deleteTodoById",
        "tags": ["Todos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Todo deleted successfully"
          }
        }
      },
      "put": {
        "summary": "Update a todo by ID",
        "operationId": "updateTodoById",
        "tags": ["Todos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/todo": {
      "post": {
        "summary": "Create a new todo",
        "operationId": "createTodo",
        "tags": ["Todos"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/todo/{id}/complete": {
      "post": {
        "summary": "Mark a todo as complete",
        "operationId": "markTodoComplete",
        "tags": ["Todos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Todo marked as complete successfully"
          }
        }
      }
    },
    "/api/todo/{id}/incomplete": {
      "post": {
        "summary": "Mark a todo as incomplete",
        "operationId": "markTodoIncomplete",
        "tags": ["Todos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Todo marked as incomplete successfully"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Tokens": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          }
        },
        "required": ["refreshToken", "accessToken"]
      },
      "TodoRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "ToDo title"
          },
          "description": {
            "type": "string",
            "example": "ToDo description"
          }
        },
        "required": ["title", "description"]
      },
      "TodoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "completed": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "example": "ToDo description"
          }
        },
        "required": ["id", "title", "userId", "completed", "createdAt", "description"]
      },
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "username1"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": ["username", "password"]
      }
    }
  }
}